Update with array_matrix crate.
Placing the guests into a matrix allowed for easy access to the
modifier values. For any guest column in the matrix there is a row
containing all the possible modifier values for the possibly adjacent
guest.

The next part is trying to implement the ABC to evaluate the possible
seating arrangement solutions.