Add solution for day 15 2015 part 1.
The solution was failing because of the subpar distribution function. It
was redone to generate random amounts for each ingredients instead of
randoming incrementing. This allowed for the generation of all of the
possible distributions (probably) and thus finding the solution.

Initial code for part 2 is using an Option type to maybe find a specific
calorie count for the solution.